Split BSTree into BSTree and BSTreeKV
